Refactor tests using LoadSceneAttribute
