+++ Tests/Runtime/Operators/TouchAndHoldOperatorTest.cs
@@ -3,31 +3,22 @@
 
 using System.Linq;
 using System.Threading.Tasks;
-using Cysharp.Threading.Tasks;
 using NUnit.Framework;
+using TestHelper.Attributes;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 using UnityEngine.TestTools;
 
 namespace TestHelper.Monkey.Operators
 {
-    [UnityPlatform(RuntimePlatform.OSXEditor, RuntimePlatform.WindowsEditor, RuntimePlatform.LinuxEditor)]
     [TestFixture]
     public class TouchAndHoldOperatorTest
     {
-        [SetUp]
-        public async Task SetUp()
-        {
-#if UNITY_EDITOR
-            await UnityEditor.SceneManagement.EditorSceneManager.LoadSceneAsyncInPlayMode(
-                "Packages/com.nowsprinting.test-helper.monkey/Tests/Scenes/Operators.unity",
-                new LoadSceneParameters(LoadSceneMode.Single));
-#endif
-        }
+        private const string TestScene = "Packages/com.nowsprinting.test-helper.monkey/Tests/Scenes/Operators.unity";
 
         [TestCase("UsingOnPointerDownUpHandler", "OnPointerDown", "OnPointerUp")]
         [TestCase("UsingPointerDownUpEventTrigger", "ReceivePointerDown", "ReceivePointerUp")]
         [TestCase("DestroyItselfIfPointerDown", "OnPointerDown", "DestroyImmediate")]
+        [LoadScene(TestScene)]
         public async Task TouchAndHold(string targetName, string expectedMessage1, string expectedMessage2)
         {
             var target = InteractiveComponentCollector.FindInteractiveComponents(false)
@@ -41,6 +32,7 @@ namespace TestHelper.Monkey.Operators
 
         [TestCase("UsingOnPointerClickHandler")]
         [TestCase("UsingPointerClickEventTrigger")]
+        [LoadScene(TestScene)]
         public void CanNotTouchAndHold(string targetName)
         {
             var target = InteractiveComponentCollector.FindInteractiveComponents(false)
